truck: rename structure folder and files
