+++ src/modules/user/services/delete-user.service.ts
@@ -1,7 +1,7 @@
 import { Inject, Injectable } from '@nestjs/common';
 import { UserRepository } from '../repository/user.repository';
-import { IUserRepository, User } from '../structure/repository.structure';
-import { UserStatus } from '../structure/user-status.enum';
+import { IUserRepository, User } from '../interfaces/repository.interface';
+import { UserStatus } from '../interfaces/user-status.enum';
 
 @Injectable()
 export class DeleteUserService {
