+++ src/modules/user/services/get-user-by-id.service.ts
@@ -1,6 +1,6 @@
 import { Inject, Injectable } from '@nestjs/common';
 import { UserRepository } from '../repository/user.repository';
-import { IUserRepository, User } from '../structure/repository.structure';
+import { IUserRepository, User } from '../interfaces/repository.interface';
 
 @Injectable()
 export class GetUserByIdService {
