+++ src/modules/user/services/user-by-filter.service.ts
@@ -1,7 +1,7 @@
 import { Inject, Injectable } from '@nestjs/common';
 import { UserRepository } from '../repository/user.repository';
-import { IUserRepository, User } from '../structure/repository.structure';
-import { IUserFilter } from '../structure/service.structure';
+import { IUserRepository, User } from '../interfaces/repository.interface';
+import { IUserFilter } from '../interfaces/user.interface';
 
 @Injectable()
 export class GetUserByFilterService {
