+++ src/modules/user/test/mocks/controller.mock.ts
@@ -1,8 +1,8 @@
 import { Request } from 'express';
 import { faker } from '@faker-js/faker';
 import { CreateUserDto } from '../../dto/create-user.dto';
-import { User } from '../../structure/repository.structure';
-import { UserStatus } from '../../structure/user-status.enum';
+import { User } from '../../interfaces/repository.interface';
+import { UserStatus } from '../../interfaces/user-status.enum';
 import { UpdateUserDto } from '../../dto/update-user.dto';
 
 export const mockFakeRequest: Request = {
