+++ src/modules/user/user.controller.ts
@@ -18,11 +18,11 @@ import { UpdateUserDto } from './dto/update-user.dto';
 import { CreateUserService } from './services/create-user.service';
 import { GetUserByIdService } from './services/get-user-by-id.service';
 import { UpdateUserService } from './services/update-user.service';
-import { User } from './structure/repository.structure';
+import { User } from './interfaces/repository.interface';
 import { DeleteUserService } from './services/delete-user.service';
 import { isPublic } from '../auth/infra/decorators/is-public.decorator';
 import { GetUserByFilterService } from './services/user-by-filter.service';
-import { IUserFilter } from './structure/service.structure';
+import { IUserFilter } from './interfaces/user.interface';
 
 @Controller('user')
 @UseFilters(new HttpExceptionFilter(new AppError()))
